created one file for all routes and imported it in app.jsJ
