--- server/routes/contactRoutes.js
@@ -1,9 +1,0 @@
-const express = require('express')
-const router = express.Router();
-const contactController = require('../controllers/contactController')
-
-// router.get('/contacts',contactController.index);
-router.post('/savecontact',contactController.saveContact);
-// router.delete('/deleteContact/:id',contactController.deleteContact);
-
-module.exports = router;
