+++ templates/pages/blog.ejs
@@ -27,7 +27,7 @@
             </div>
         </div>
     </main>
-    <%- include('pages/partials/footer.ejs') %>
+    <%- include('partials/footer.ejs') %>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase-app.js"></script>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase-firestore.js"></script>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase-auth.js"></script>
