+++ templates/pages/blogSingle.html
@@ -40,7 +40,7 @@
             </div>
         </div>
     </main>
-    <%- include('pages/partials/footer.ejs') %>
+    <%- include('partials/footer.ejs') %>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase-app.js"></script>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase.js"></script>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase-firestore.js"></script>
