+++ templates/pages/contact.ejs
@@ -49,7 +49,7 @@
            </form>
        </div> 
     </main>
-    <%- include('pages/partials/footer.ejs') %>
+    <%- include('partials/footer.ejs') %>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase-app.js"></script>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase-firestore.js"></script>
     <script src="https://www.gstatic.com/firebasejs/7.19.1/firebase-auth.js"></script>
